i am in pain
